Agent/requirements.txt
torch
torchvision
pillow

pandas
openpyxl
requests
numpy

fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart
3